+++ app/controllers/categories_controller.rb
@@ -65,7 +65,7 @@ class CategoriesController < ApplicationController
   end
 
   def category_params
-    params.require(:category).permit(:name, :slug)
+    params.require(:category).permit(:name, :slug, :announcement)
   end
 
   def scope_order(scope)