+++ db/migrate/20200919141201_add_announcement_to_categories.rb
@@ -1,0 +1,6 @@
+class AddAnnouncementToCategories < ActiveRecord::Migration[6.0]
+  def change
+    add_column :categories, :announcement, :boolean, null: false, default: false
+    add_index :categories, :announcement
+  end
+end